Ajout de la classe DefaultTheme, affichage du score durant le jeu et durant le game over, modification et ajouts d'images

+++ core/src/com/pts4/game/Score.java
@@ -33,7 +33,7 @@ public class Score {
         return temp;
     }
 
-    public float getScoreMax(){
+    public int getScoreMax(){
         return this.scoreMax;
     }
